Tidy up python engine
